refactor: modularize features
