+++ lib/features/chat/data/gemini_service.dart
