+++ lib/features/note/domain/entities/note.dart
