+++ lib/features/note/presentation/screens/note_list_for_day_screen.dart
@@ -3,13 +3,13 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
-import '../models/note.dart';
+import '../../domain/entities/note.dart';
 import '../providers/note_provider.dart';
-import '../services/auth_service.dart';
+import 'package:notes_reminder_app/services/auth_service.dart';
 import 'note_detail_screen.dart';
-import '../pandora_ui/hint_chip.dart';
-import '../widgets/note_card.dart';
-import '../widgets/route_transitions.dart';
+import 'package:notes_reminder_app/pandora_ui/hint_chip.dart';
+import 'package:notes_reminder_app/widgets/note_card.dart';
+import 'package:notes_reminder_app/widgets/route_transitions.dart';
 
 class NoteListForDayScreen extends StatelessWidget {
   final DateTime date;
