+++ lib/features/note/presentation/screens/voice_to_note_screen.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:speech_to_text/speech_to_text.dart' as stt;
 
 import '../providers/note_provider.dart';
-import '../services/gemini_service.dart';
+import 'package:notes_reminder_app/features/chat/domain/usecases/chat_with_gemini.dart';
 
 class VoiceToNoteScreen extends StatefulWidget {
   final stt.SpeechToText speech;
@@ -26,6 +26,7 @@ class _VoiceToNoteScreenState extends State<VoiceToNoteScreen> {
   bool _isProcessing = false;
   double _level = 0;
   double _maxLevel = 1;
+  final ChatWithGemini _chatWithGemini = ChatWithGemini();
 
   @override
   void initState() {
@@ -90,7 +91,7 @@ class _VoiceToNoteScreenState extends State<VoiceToNoteScreen> {
     final prompt = AppLocalizations.of(context)!
         .convertSpeechPrompt(_recognized);
     final l10n = AppLocalizations.of(context)!;
-    final reply = await GeminiService().chat(prompt, l10n);
+    final reply = await _chatWithGemini(prompt, l10n);
     if (!mounted) return;
     context.read<NoteProvider>().setDraft(reply);
     setState(() => _isProcessing = false);
