+++ lib/widgets/tag_filtered_notes_list.dart
@@ -3,9 +3,9 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
-import '../models/note.dart';
-import '../providers/note_provider.dart';
-import '../screens/note_list_for_day_screen.dart';
+import 'package:notes_reminder_app/features/note/domain/entities/note.dart';
+import 'package:notes_reminder_app/features/note/presentation/providers/note_provider.dart';
+import 'package:notes_reminder_app/features/note/presentation/screens/note_list_for_day_screen.dart';
 import 'notes_list.dart';
 import 'tag_filter_menu.dart';
 import 'route_transitions.dart';
