+++ test/note_search_delegate_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:notes_reminder_app/models/note.dart';
-import 'package:notes_reminder_app/screens/note_search_delegate.dart';
+import 'package:notes_reminder_app/features/note/domain/entities/note.dart';
+import 'package:notes_reminder_app/features/note/presentation/screens/note_search_delegate.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 void main() {
