+++ test/notes_list_unsynced_test.dart
@@ -5,10 +5,10 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'package:notes_reminder_app/providers/note_provider.dart';
+import 'package:notes_reminder_app/features/note/presentation/providers/note_provider.dart';
 import 'package:notes_reminder_app/widgets/notes_list.dart';
-import 'package:notes_reminder_app/models/note.dart';
-import 'package:notes_reminder_app/services/note_repository.dart';
+import 'package:notes_reminder_app/features/note/domain/entities/note.dart';
+import 'package:notes_reminder_app/features/note/data/note_repository.dart';
 
 class MockRepo extends Mock implements NoteRepository {}
