+++ test/tag_filtered_notes_list_test.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:notes_reminder_app/providers/note_provider.dart';
+import 'package:notes_reminder_app/features/note/presentation/providers/note_provider.dart';
 import 'package:notes_reminder_app/widgets/tag_filtered_notes_list.dart';
-import 'package:notes_reminder_app/models/note.dart';
+import 'package:notes_reminder_app/features/note/domain/entities/note.dart';
 
 void main() {
   testWidgets('filters notes by tag', (tester) async {
