+++ test/voice_to_note_screen_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:speech_to_text/speech_to_text.dart' as stt;
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:notes_reminder_app/screens/voice_to_note_screen.dart';
+import 'package:notes_reminder_app/features/note/presentation/screens/voice_to_note_screen.dart';
 
 class MockSpeechToText extends Mock implements stt.SpeechToText {}
